requirements.txt
matplotlib
numpy
requests
scipy
pillow
